requirements.txt
Adafruit_Python_DHT==1.4.0
gpiozero==1.5.1
pandas==1.0.3
picamera==1.13
RPi.GPIO==0.7.0
